add optparse and cucumber help action feature
 * need usage to work
 * need first test to work
 * need cucumber to function
 * need rspec working
 * need env.rb working
