+++ features/001_basic_usage.feature
@@ -1,0 +1,8 @@
+Feature: Basic Usage
+  Because Everyone wants a stable tool
+  The typical command line options should work
+
+  Scenario: bnzops help
+    Given I reset my options
+    And I add action 'help'
+    Then My output should contain 'Usage'
