+++ lib/bnzops.rb
@@ -1,7 +1,8 @@
 
+$BNZ_DEBUG = false
 module BNZOps end
 module BNZOps::Contrib end
 module BNZOps::Contrib::Packages end
 
-require_relative 'bnzops/cli.rb'
+require_relative 'bnzops/action.rb'
