+++ lib/bnzops/action.rb
@@ -1,0 +1,4 @@
+module BNZOps::Action end
+
+require_relative 'action/configure.rb'
+
